// src/api/readingHistory-service/models/page-record-response.ts
/* tslint:disable */
/* eslint-disable */
/**
 * OpenAPI definition
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import type { PageableObject } from './pageable-object';
// May contain unused imports in some cases
// @ts-ignore
import type { RecordResponse } from './record-response';
// May contain unused imports in some cases
// @ts-ignore
import type { SortObject } from './sort-object';

/**
 * 
 * @export
 * @interface PageRecordResponse
 */
export interface PageRecordResponse {
    /**
     * 
     * @type {number}
     * @memberof PageRecordResponse
     */
    'totalPages'?: number;
    /**
     * 
     * @type {number}
     * @memberof PageRecordResponse
     */
    'totalElements'?: number;
    /**
     * 
     * @type {number}
     * @memberof PageRecordResponse
     */
    'size'?: number;
    /**
     * 
     * @type {Array<RecordResponse>}
     * @memberof PageRecordResponse
     */
    'content'?: Array<RecordResponse>;
    /**
     * 
     * @type {number}
     * @memberof PageRecordResponse
     */
    'number'?: number;
    /**
     * 
     * @type {SortObject}
     * @memberof PageRecordResponse
     */
    'sort'?: SortObject;
    /**
     * 
     * @type {boolean}
     * @memberof PageRecordResponse
     */
    'first'?: boolean;
    /**
     * 
     * @type {boolean}
     * @memberof PageRecordResponse
     */
    'last'?: boolean;
    /**
     * 
     * @type {number}
     * @memberof PageRecordResponse
     */
    'numberOfElements'?: number;
    /**
     * 
     * @type {PageableObject}
     * @memberof PageRecordResponse
     */
    'pageable'?: PageableObject;
    /**
     * 
     * @type {boolean}
     * @memberof PageRecordResponse
     */
    'empty'?: boolean;
}