// src/api/readingHistory-service/models/record-response.ts
/* tslint:disable */
/* eslint-disable */
/**
 * OpenAPI definition
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 
 * @export
 * @interface RecordResponse
 */
export interface RecordResponse {
    /**
     * 
     * @type {string}
     * @memberof RecordResponse
     */
    'userId'?: string;
    /**
     * 
     * @type {string}
     * @memberof RecordResponse
     */
    'chapterId'?: string;
    /**
     * 
     * @type {string}
     * @memberof RecordResponse
     */
    'bookId'?: string;
    /**
     * 
     * @type {string}
     * @memberof RecordResponse
     */
    'lastReadAt'?: string;
}